[API] Factored out embedded initializers to Intermodal::Rails::Rails4Stack
There are some additional things like can be refactored. For example,
a more sensible place to put in Warden. Maybe in the
the Intermodal::Rails::UseWarden concern -- or better yet, rename that
to Intermodal::Rails::UseXAuthToken

Will punt this for the time when the internal rspec is working again
instead of relying on an external API app to test Intermodal against.